+++ src/components/layout/Layout.tsx
@@ -4,8 +4,8 @@ import React from 'react'
 import Helmet from 'react-helmet'
 import { Option } from 'space-lift'
 
+import { PageAttributes } from '~/types'
 import { Node } from '~/types/graph'
-import { PageAttributes } from '~/types/page'
 import { Footer } from '..'
 import { Header } from '../header'
 import * as style from './style'
