--- src/types/BackgroundImageAttribute.ts
@@ -1,8 +1,0 @@
-import { Fluid, GatsbyImage } from './image'
-export type BackgroundImageAttribute =
-  | (GatsbyImage<Fluid> & {
-      original: {
-        src: string
-      }
-    })
-  | string
