--- src/types/widget.ts
@@ -1,10 +1,0 @@
-export type CTA = {
-  palette: 'primary' | 'secondary'
-  label: string
-}
-
-export type Image = {
-  src: string
-  hiddenHeaderContent?: string
-  hiddenHeaderLevel?: number
-}
